Comment regridding control script
This moves many descriptions from the readme file into the control
script, making it easier to adapt. Also, the paths are adapted to my
computer.

In the long term, this should be done differently (i.e. the possibility
of name parsing, etc.), but for now let's just get it to work :)