Convert output values to byte integer
Before, we were working with byte values of normal length, so the
conversion to byte worked fine. The higher resolution data uses 16-bit
integers, so when the value was converted to normal byte the initial
number was too small.

The reason was that the first division gave an integer value back (in
this example 255/65536). Thta value was 0, and when it was multiplied
with the observed value it of course also produced a 0.

The solution: Divide with a real value and then multiply with the
observation, thus converting the number into the right number range. The
result was then rounded to the next integer. Now the regridded values
look fine.