Bugfix: Correct loop entry requirements
In magicsol-v0goes.c a loop was remaining unentered because one of its
requirements was uninitialized on the first run. Rather than setting the
variable to 0 during initialization, gcc 4.7 allocates the memory it
needs. If the variable is later read without that memory address being
specifically set, the binary code stored at that address is interpreted
as a value of the variable's type. Therefore the value of anzahl was
"random" and normally above the entry requirements for the loop.

This led to the files not being read and, at the end, an underflow with
no results.

The fix was to set the value to 0 at initialization. Older compilers had
done this, but the newer didn't, so the program could only be compiled
on computers using older versions.