Add allsky code for irradiation with GOES
The new code is adapted from the old code, which is still stored in the allsky
folder. The current example script, magic_daniel.sh, calculates global and
direct irradiation from cloud albedo data. Remember to adapt it to the paths in
concern. Dimensions, etc. are in the header of each file.

I plan on documenting this better in the future, but for now, if you're looking
for the meaning of the words in the config file that magic.exe uses, check out
"allsky_goes/magic-config.inp - beispiel fuer goes-e".

The sources for the magic executable are found in the allsky_goes/c. The
executable should be run in the allsky, with the climatologies and look up
tables (luts) as subfolders. The output folder can be located wherever you like;
for this example, I've also put it in the allsky_goes folder, but when I run it
on the cluster it'll of course be located elsewhere.